requirements.txt
attrs
twisted
autobahn
sqlalchemy
passlib
db-dumper
klein
emote-utils
